+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Helper types `ModelDataType<M>` and `ModelInitType<M>` which resolve to the
+  model `M`'s data-type and init-type.
 ### Fixed
 - Do not mangle `id` property of model when minifying the React adapter.